version 5

+++ frontoffice/pedido.php
@@ -36,14 +36,22 @@
 	<th>Cantidad</th>
 	<th>Creado</th>
 	<th>Categoria</th>
+<?php 
+	if($_SESSION['rol']== 'administrador'){
+		// Para agregar el contenido
+ ?>
 	<th>Eliminar</th>
 	<th>Modificar</th>
+<?php 
+	}
+ ?>
 	<th>Seleccionar</th>
 	</tr>
 
 <?php 
 		
 
+		
 		if(isset($_REQUEST['id'])){
 			$id = $_REQUEST['id'];
 			$verproductos = "SELECT * FROM productos JOIN categorias_productos USING (idproducto)
@@ -59,13 +67,28 @@
 					echo "<td>".$registro['precio']."</td>";
 					echo "<td>".$registro['marca']."</td>";
 					echo "<td><img id='imagen' src='".$registro['ruta']."'alt='imagen'></img></td>";
-					echo "<td><input name='canselec' value='".$registro['cantidad']."'></td>";
+				
+					echo "<td>";
+					$cantidad = $registro['cantidad'];
+					if($cantidad !=0){
+						echo "<select name='cantidad'>";
+						for ($i=1; $i <= $cantidad ; $i++) { 
+							echo "<option value='$i'>$i</option>";
+						}
+						echo "</select>";
+					}else{
+						echo "<h2>No hay productos por ahora</h2>";
+					}
+					echo "</td>";
 					echo "<td>".$registro['createdAt']."</td>";
 					echo "<td>".$registro['nombre']."</td>";
-					echo "<td>"."<button><a href='index.php?sec=eliminarpro&id=".$id."'>Eliminar</a></button>"."</td>";
-					echo "<td><button><a href='index.php?sec=modificarpro&cant=".$registro['cantidad']."&ctg=".$registro['nombre']."&m=".$registro['marca']."&id=".$id."&t=".$registro['titulo']."&p=".$registro['precio']."&d=".$registro['descripcion']."&img=".$registro['ruta']."'>Modificar</a></button></td>";
+					if($_SESSION['rol']== 'administrador'){
+						echo "<td>"."<button><a href='index.php?sec=eliminarpro&id=".$id."'>Eliminar</a></button>"."</td>";
+						echo "<td><button><a href='index.php?sec=modificarpro&cant=".$registro['cantidad']."&ctg=".$registro['nombre']."&m=".$registro['marca']."&id=".$id."&t=".$registro['titulo']."&p=".$registro['precio']."&d=".$registro['descripcion']."&img=".$registro['ruta']."'>Modificar</a></button></td>";
+					}
 					echo "<td><button><a href='index.php?sec=pedido&id=".$id."'>Seleccionar</a></button></td>";
 					echo "</tr>";	
+				
 				}
 		}
